phase-test/requirements.txt
langchain-groq
langchain-experimental
langchain-community
xgboost
plotly
pandas==2.2.2
requests==2.32.4
numpy<2.4.0
scikit-learn
streamlit
pyngrok
opencv-python-headless
pypdf
tabulate
python-dotenv
